GoNexus-AI/requirements.txt
grpcio
grpcio-tools
openai
chromadb
python-dotenv
sentence-transformers